Refactor App.tsx to contain the logic
